requirements.txt
Django
scikit-learn
djangorestframework
pandas